requirements.txt
pytest
PySide2
requests
pyyaml
